fix: consolidate StoredAnalysis type
- Moved StoredAnalysis type to shared types directory
- Updated storage utility to use consolidated type
- Fixed type imports across components
- Added type re-export for backward compatibility
