+++ components/contract-analyzer/types/storage.ts
@@ -1,15 +1,4 @@
-import { AnalysisResult } from './analysis';
+// This file is deprecated.
+// Import types from @/types/storage instead.
 
-export interface StorageOptions {
-  maxItems?: number;
-  storageKey?: string;
-}
-
-export interface StoredAnalysis {
-  id: string;
-  fileName: string;
-  fileHash: string;  // For content-based comparison
-  fileSize: number;  // Quick initial check before hash comparison
-  analysis: AnalysisResult;
-  analyzedAt: string;
-}
+export * from '@/types/storage';
